CHAP14: Remove call to @requireJs to avoid problem in production.
The call to @helper.requireJs was causing the wrong URL to be entered for
main.js in production: the javascripts folder was replaced by javascripts-min,
which wasn't actually where main.js was.

Hard-coding the correct URL fares better, though this is hardly a great
solution. Anyway, at least the current version deploys.